refactor project callback func, dep field func
