+++ apps/mappings/imports/modules/projects.py
@@ -1,8 +1,16 @@
+import logging
 from datetime import datetime
-from typing import List
+from typing import List, Dict
 from apps.mappings.imports.modules.base import Base
 from apps.sage300.models import CostCategory
-from fyle_accounting_mappings.models import DestinationAttribute
+from fyle_accounting_mappings.models import DestinationAttribute, ExpenseAttribute
+from fyle_integrations_platform_connector import PlatformConnector
+from apps.sage300.dependent_fields import update_and_disable_cost_code
+from apps.mappings.helpers import prepend_code_to_name
+from apps.workspaces.models import FyleCredential, ImportSetting
+
+logger = logging.getLogger(__name__)
+logger.level = logging.INFO
 
 
 class Project(Base):
@@ -67,3 +75,76 @@ class Project(Base):
                     payload.append(project)
 
         return payload
+
+
+def disable_projects(workspace_id: int, projects_to_disable: Dict, *args, **kwargs):
+    """
+    Disable projects in Fyle when the projects are updated in Sage 300.
+    This is a callback function that is triggered from accounting_mappings.
+    projects_to_disable object format:
+    {
+        'destination_id': {
+            'value': 'old_project_name',
+            'updated_value': 'new_project_name',
+            'code': 'old_project_code',
+            'updated_code': 'new_project_code'
+        }
+    }
+
+    """
+    fyle_credentials = FyleCredential.objects.get(workspace_id=workspace_id)
+    platform = PlatformConnector(fyle_credentials=fyle_credentials)
+    platform.projects.sync()
+
+    use_code_in_naming = ImportSetting.objects.filter(
+        workspace_id = workspace_id,
+        import_code_fields__contains=['JOB']
+    ).first()
+
+    project_values = []
+    for projects_map in projects_to_disable.values():
+        project_name = prepend_code_to_name(prepend_code_in_name=use_code_in_naming, value=projects_map['value'], code=projects_map['code'])
+        project_values.append(project_name)
+
+    filters = {
+        'workspace_id': workspace_id,
+        'attribute_type': 'PROJECT',
+        'value__in': project_values,
+        'active': True
+    }
+
+    # Expense attribute value map is as follows: {old_project_name: destination_id}
+    expense_attribute_value_map = {}
+    for k, v in projects_to_disable.items():
+        project_name = prepend_code_to_name(prepend_code_in_name=use_code_in_naming, value=v['value'], code=v['code'])
+        expense_attribute_value_map[project_name] = k
+
+    expense_attributes = ExpenseAttribute.objects.filter(**filters)
+
+    bulk_payload = []
+    for expense_attribute in expense_attributes:
+        code = expense_attribute_value_map.get(expense_attribute.value, None)
+        if code:
+            payload = {
+                'name': expense_attribute.value,
+                'code': code,
+                'description': 'Sage 300 Project - {0}, Id - {1}'.format(
+                    expense_attribute.value,
+                    code
+                ),
+                'is_enabled': False,
+                'id': expense_attribute.source_id
+            }
+            bulk_payload.append(payload)
+        else:
+            logger.error(f"Project with value {expense_attribute.value} not found | WORKSPACE_ID: {workspace_id}")
+
+    if bulk_payload:
+        logger.info(f"Disabling Projects in Fyle | WORKSPACE_ID: {workspace_id} | COUNT: {len(bulk_payload)}")
+        platform.projects.post_bulk(bulk_payload)
+    else:
+        logger.info(f"No Projects to Disable in Fyle | WORKSPACE_ID: {workspace_id}")
+    print("reached here")
+    update_and_disable_cost_code(workspace_id, projects_to_disable, platform, use_code_in_naming)
+    platform.projects.sync()
+    return bulk_payload
