+++ apps/sage300/helpers.py
@@ -1,18 +1,8 @@
-
-from datetime import datetime, timezone
 import logging
+from datetime import datetime, timezone
 
-from typing import Dict
 from django.utils.module_loading import import_string
-
-from apps.workspaces.models import Workspace, Sage300Credential, FyleCredential, ImportSetting
-from fyle_accounting_mappings.models import ExpenseAttribute
-from fyle_integrations_platform_connector import PlatformConnector
-from apps.sage300.models import CostCategory
-from apps.fyle.models import DependentFieldSetting
-from apps.sage300.dependent_fields import post_dependent_cost_code
-from apps.mappings.models import ImportLog
-from apps.mappings.helpers import prepend_code_to_name
+from apps.workspaces.models import Workspace, Sage300Credential
 
 logger = logging.getLogger(__name__)
 logger.level = logging.INFO
@@ -65,114 +55,3 @@ def sync_dimensions(sage300_credential: Sage300Credential, workspace_id: int) ->
         except Exception as exception:
             # Log any exceptions that occur during synchronization
             logger.info(exception)
-
-
-def disable_projects(workspace_id: int, projects_to_disable: Dict, *args, **kwargs):
-    """
-    Disable projects in Fyle when the projects are updated in Sage 300.
-    This is a callback function that is triggered from accounting_mappings.
-    projects_to_disable object format:
-    {
-        'destination_id': {
-            'value': 'old_project_name',
-            'updated_value': 'new_project_name',
-            'code': 'old_project_code',
-            'updated_code': 'new_project_code'
-        }
-    }
-
-    """
-    fyle_credentials = FyleCredential.objects.get(workspace_id=workspace_id)
-    platform = PlatformConnector(fyle_credentials=fyle_credentials)
-    platform.projects.sync()
-
-    use_code_in_naming = ImportSetting.objects.filter(
-        workspace_id = workspace_id,
-        import_code_fields__contains=['JOB']
-    ).first()
-
-    project_values = []
-    for projects_map in projects_to_disable.values():
-        project_name = prepend_code_to_name(prepend_code_in_name=use_code_in_naming, value=projects_map['value'], code=projects_map['code'])
-        project_values.append(project_name)
-
-    filters = {
-        'workspace_id': workspace_id,
-        'attribute_type': 'PROJECT',
-        'value__in': project_values,
-        'active': True
-    }
-
-    # Expense attribute value map is as follows: {old_project_name: destination_id}
-    expense_attribute_value_map = {}
-    for k, v in projects_to_disable.items():
-        project_name = prepend_code_to_name(prepend_code_in_name=use_code_in_naming, value=v['value'], code=v['code'])
-        expense_attribute_value_map[project_name] = k
-
-    expense_attributes = ExpenseAttribute.objects.filter(**filters)
-
-    bulk_payload = []
-    for expense_attribute in expense_attributes:
-        code = expense_attribute_value_map.get(expense_attribute.value, None)
-        if code:
-            payload = {
-                'name': expense_attribute.value,
-                'code': code,
-                'description': 'Sage 300 Project - {0}, Id - {1}'.format(
-                    expense_attribute.value,
-                    code
-                ),
-                'is_enabled': False,
-                'id': expense_attribute.source_id
-            }
-            bulk_payload.append(payload)
-        else:
-            logger.error(f"Project with value {expense_attribute.value} not found | WORKSPACE_ID: {workspace_id}")
-
-    if bulk_payload:
-        logger.info(f"Disabling Projects in Fyle | WORKSPACE_ID: {workspace_id} | COUNT: {len(bulk_payload)}")
-        platform.projects.post_bulk(bulk_payload)
-    else:
-        logger.info(f"No Projects to Disable in Fyle | WORKSPACE_ID: {workspace_id}")
-
-    update_and_disable_cost_code(workspace_id, projects_to_disable, platform, use_code_in_naming)
-    platform.projects.sync()
-    return bulk_payload
-
-
-def update_and_disable_cost_code(workspace_id: int, cost_codes_to_disable: Dict, platform: PlatformConnector, use_code_in_naming: bool):
-    """
-    Update the job_name in CostCategory and disable the old cost code in Fyle
-    """
-    dependent_field_setting = DependentFieldSetting.objects.filter(is_import_enabled=True, workspace_id=workspace_id).first()
-
-    if dependent_field_setting:
-        filters = {
-            'job_id__in':list(cost_codes_to_disable.keys()),
-            'workspace_id': workspace_id
-        }
-        cost_code_import_log = ImportLog.create('COST_CODE', workspace_id)
-        # This call will disable the cost codes in Fyle that has old project name
-        posted_cost_codes = post_dependent_cost_code(cost_code_import_log, dependent_field_setting, platform, filters, is_enabled=False)
-
-        logger.info(f"Disabled Cost Codes in Fyle | WORKSPACE_ID: {workspace_id} | COUNT: {len(posted_cost_codes)}")
-
-        # here we are updating the CostCategory with the new project name
-        bulk_update_payload = []
-        for destination_id, value in cost_codes_to_disable.items():
-            updated_job_name = prepend_code_to_name(prepend_code_in_name=use_code_in_naming, value=value['updated_value'], code=value['updated_code'])
-
-            cost_categories = CostCategory.objects.filter(
-                workspace_id=workspace_id,
-                job_id=destination_id
-            ).exclude(job_name=updated_job_name)
-
-            for cost_category in cost_categories:
-                cost_category.job_code = value['updated_code']
-                cost_category.job_name = value['updated_value']
-                cost_category.updated_at = datetime.now(timezone.utc)
-                bulk_update_payload.append(cost_category)
-
-        if bulk_update_payload:
-            logger.info(f"Updating Cost Categories | WORKSPACE_ID: {workspace_id} | COUNT: {len(bulk_update_payload)}")
-            CostCategory.objects.bulk_update(bulk_update_payload, ['job_name', 'job_code', 'updated_at'], batch_size=50)
